REFACTOR: move class FittedText to pytools.viz.util
